Move ResourceLock to submodule.
